Move common parse mentions logic unto busniess layer
